requirements.txt
streamlit
textblob
Pillow
numpy
matplotlib
tensorflow
scikit-learn
random